🗑️: domain.Composerの削除
modelに統合
